requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.11
pytest==8.4.2
requests==2.32.3
python-decouple==3.8
flake8==7.0.0
black==24.2.0
